Fix bet to back

+++ app/src/main/java/com/kvad/totalizator/betfeature/BetRequest.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class BetRequest(
     @SerializedName("event_id") val eventId: String,
     @SerializedName("choice") val choice: String,
-    @SerializedName("amount") val amount: Double
+    @SerializedName("amount") val amount: Int
 )
